requirement.txt
Click==7.0
configparser==3.7.4
Fabric3==1.14.post1
jenkinsapi==0.3.8
logger==1.4
lxml==4.3.3
paramiko==2.4.2
python-jenkins==1.4.0
requests==2.21.0
six==1.12.0
urllib3==1.24.1
wait-for==1.0.11
